LayoutTests/fast/dom/css-inline-style-important.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../js/resources/js-test-pre.js"></script>
<style type="text/css">
#e { color:red; visibility: hidden;}
</style>
</head>
<body>
<p id="e">This is the test element.</p>
<script>

description("This test checks to see whether setting a CSS value including !important on an inline style property fails (as it should).");

var e = document.getElementById('e');
e.style.color = "green !important";

shouldBe('window.getComputedStyle(e).color', '"rgb(255, 0, 0)"');

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
